# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkOSWebhookResponse": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          }
        },
        "required": [
          "accepted"
        ],
        "title": "WorkOSWebhookResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "CSRFHeader": {
        "description": "Signed double-submit CSRF token required for unsafe browser requests in WorkOS auth mode.",
        "in": "header",
        "name": "x-archagent-csrf",
        "type": "apiKey"
      },
      "CloudTasksOIDC": {
        "description": "Cloud Run OIDC token presented by Google Cloud Tasks in production.",
        "scheme": "bearer",
        "type": "http"
      },
      "CollectorBearer": {
        "description": "Collector credential returned by /collector/v1/register or /collector/v1/credentials/rotate.",
        "scheme": "bearer",
        "type": "http"
      },
      "DevelopmentIdentityHeader": {
        "description": "Local development/test identity selector. Disabled in production.",
        "in": "header",
        "name": "x-archagent-user",
        "type": "apiKey"
      },
      "WorkOSSessionCookie": {
        "description": "Sealed WorkOS AuthKit session cookie used in hosted auth mode.",
        "in": "cookie",
        "name": "__Host-archagent-session",
        "type": "apiKey"
      },
      "WorkOSWebhookSignature": {
        "description": "WorkOS webhook signature header validated with the configured webhook secret.",
        "in": "header",
        "name": "workos-signature",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "\nArchAgent API contract for browser frontend, collector, staff, webhook, and worker integrations.\n\nUse `/openapi.json` as the machine-readable source of truth. Swagger UI is available at `/docs`\nand ReDoc is available at `/redoc` when the FastAPI app is running.\n",
    "title": "ArchAgent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/webhooks/workos/{path_token}": {
      "post": {
        "description": "Validate a signed WorkOS webhook, persist its receipt idempotently, and enqueue asynchronous synchronization.",
        "operationId": "receive_workos_webhook_webhooks_workos__path_token__post",
        "parameters": [
          {
            "in": "path",
            "name": "path_token",
            "required": true,
            "schema": {
              "title": "Path Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOSWebhookResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "WorkOSWebhookSignature": []
          }
        ],
        "summary": "Receive WorkOS Webhook",
        "tags": [
          "Webhooks"
        ]
      }
    }
  },
...
}